+++ src/twitch/command/CommandActionEventNode.cpp
@@ -446,9 +446,9 @@ bool CommandActionEventNode::initEventNode(const std::string& labelText, CCObjec
         if (node.label == labelText) {
             m_eventDescription = node.description;
             break;
-        };
-    };
-
+        }
+    }
+    
     auto labelNode = CCNode::create();
     labelNode->setPosition(0, 0);
 
@@ -484,26 +484,26 @@ std::vector<EventNodeInfo> CommandActionEventNode::getAllEventNodes() {
         {"kill_player", "Destroy Player", "Destroy player. Self-explanatory. <cr>Don't use this while beating extremes!</c>"},
         {"jump", "Jump", "Force the player to jump. You can set it to also hold jump."},
         {"move", "Move Player", "Move the player left or right. Lets you pick the player, direction and the distance to move."},
-        {"color_player", "Color Player", "Set the player's color based on the RGB value. <cr>Broken on Android users at this moment.</c>."},
+        {"color_player", "Color Player", "Set the player's color based on the RGB value. <cr>Broken on Android users at this moment.</c>"},
         {"wait", "Wait", "Pauses the command sequence for a set amount of time (in seconds). <cg>Use as a delay between actions.</c>"},
         {"notification", "Notification", "Shows a notification message on the screen. <cg>Supports the use of identifiers.</c>."},
         {"alert_popup", "Alert Popup", "Shows an alert popup like this one you reading. <cg>Supports the use of identifiers.</c>."},
-        {"keycode", "Key Code", "Simulates a key press or release. <cr>Does not work on mobile users</c>."},
-        {"profile", "Profile", "Opens the Player Profile in-game. <cg>Only accepts Account ID only!</c>."},
+        {"keycode", "Key Code", "Simulates a key press or release. <cr>Does not work on mobile users</c>"},
+        {"profile", "Profile", "Opens the Player Profile in-game. <cg>Only accepts Account ID only!</c>"},
         {"nothing", "Nothing", "Does nothing at all."},
-        {"scale_player", "Scale Player", "Scales the player in-game. <cr>Does not affect the player hitbox.</c>."},
-        {"sound", "Sound Effect", "Plays a sound effect. <cg>Only uses the provided sound effect included in game.</c>."},
-        {"stop_all_sounds", "Stop All Sounds", "Stops all currently playing sound effects immediately."},
-        {"gravity", "Set Gravity", "Sets the player's gravity to a specified value for a duration. <cg>Use to modify gravity temporarily.</c>."},
-        {"speed_player", "Speed Player", "Sets the player's speed to a specified value for a duration. <cg>Use to modify speed temporarily.</c>."},
+        {"scale_player", "Scale Player", "Scales the player in-game. <cr>Does not affect the player hitbox.</c>"},
+        {"sound", "Sound Effect", "Plays a sound effect. <cg>Only uses the provided sound effect included in game.</c>"},
+        {"stop_all_sounds", "Stop All Sounds", "Stops all currently playing sound effects immediately"},
+        {"gravity", "Gravity Player", "Sets the player's gravity to a specified value for a duration."},
+        {"speed_player", "Speed Player", "Sets the player's speed to a specified value for a duration."},
         {"restart_level", "Restart Level", "Restarts the entire level."},
-        {"noclip", "Noclip", "Enables or disables noclip mode for the player."}
+        {"noclip", "Noclip", "Enables or disables noclip mode for the player. <cr>This does not have Safe Mode, use with caution!</c> <cy>Disables upon exiting the level</c>"}
     };
 
     auto mod = Mod::get();
     if (mod && mod->getSettingValue<bool>("experimental"))
         nodes.push_back(
-            { "edit_camera", "Edit Camera", "Edit Camera's Skew, Rotation, and Scale." }
+            { "edit_camera", "Edit Camera", "Edit Camera's Skew, Rotation, and Scale. <cy>Experimental Feature. May crash your game.</c>" }
         );
 
     return nodes;